create package for toolkit, update readmes
