+++ .gitignore
@@ -1,9 +1,1 @@
 node_modules
-commands/init/node_modules
-commands/init/build
-commands/init/simulator/massa-sc-tester
-commands/init/simulator/massa-sc-tester.exe
-commands/init/simulator/ledger.json
-commands/init/simulator/trace.json
-commands/init/.env
-
