+++ packages/sc-compiler/README.md
@@ -1,0 +1,3 @@
+# Massa smart-contract Compiler
+
+This package can compile all the smart contracts in `assembly/contracts` with the command `npx compile`.
