+++ packages/sc-deployer/README.md
@@ -1,0 +1,3 @@
+# Massa smart-contract Deployer
+
+This package can deploy a smart contract with an initialization function.
