+++ packages/sc-toolkit/.gitignore
@@ -1,0 +1,6 @@
+commands/init/build
+commands/init/simulator/massa-sc-tester
+commands/init/simulator/massa-sc-tester.exe
+commands/init/simulator/ledger.json
+commands/init/simulator/trace.json
+commands/init/.env
