+++ packages/sc-toolkit/README.md
@@ -1,0 +1,21 @@
+# massa-sc-toolkit
+
+This toolkit is meant to facilitate smart contract development.
+
+> **PREREQUISITES:**
+>
+> - NPM installed on your computer
+
+## Repository Initialization
+
+Simply run the following command:
+
+```shell
+npx @massalabs/massa-sc-toolkit init <projectName>
+```
+
+Toolkit documentation is available [here](commands/init/README.md)
+
+## Develop
+
+You can initialize a project with `npx toolkit init testDir`.
