+++ packages/sc-toolkit/commands/init/.env.example
