+++ packages/sc-toolkit/commands/init/.eslintignore
