+++ packages/sc-toolkit/commands/init/.eslintrc.json
