+++ packages/sc-toolkit/commands/init/README.md
