+++ packages/sc-toolkit/commands/init/as-pect.asconfig.json
