+++ packages/sc-toolkit/commands/init/as-pect.config.js
