+++ packages/sc-toolkit/commands/init/asconfig.json
