+++ packages/sc-toolkit/commands/init/assembly/__tests__/as-pect.d.ts
