+++ packages/sc-toolkit/commands/init/assembly/__tests__/massa-example.spec.ts
