+++ packages/sc-toolkit/commands/init/assembly/contracts/main.ts
