+++ packages/sc-toolkit/commands/init/assembly/tsconfig.json
