+++ packages/sc-toolkit/commands/init/gitignore
