+++ packages/sc-toolkit/commands/init/package.json
