+++ packages/sc-toolkit/commands/init/tsconfig.json
