+++ packages/sc-toolkit/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@massalabs/sc-toolkit",
+  "version": "0.1.0",
+  "description": "",
+  "main": "src/cli.js",
+  "type": "module",
+  "scripts": {
+    "prettier": "prettier '**/src/**/*.js' --check && as-prettier --check assembly",
+    "prettier:fix": "prettier '**/src/**/*.js' --write && as-prettier --write assembly",
+    "lint": "",
+    "lint:fix": "",
+    "fmt": "npm run prettier:fix && npm run lint:fix",
+    "fmt:check": "npm run prettier && npm run lint"
+  },
+  "dependencies": {
+    "fs-extra": "^10.1.0",
+    "yargs": "^17.6.0"
+  },
+  "author": "",
+  "license": "ISC",    
+  "bin": {
+    "toolkit": "src/cli.js"
+  }
+}
